+++ .github/workflows/publish.yml
@@ -14,10 +14,11 @@ jobs:
       with:
         node-version: '20'
         registry-url: 'https://registry.npmjs.org/'
-    - run: npm i
-    - run: npm run lint
-    - run: npm run build
-    - run: npm run test
+    - run: corepack enable
+    - run: yarn config set enableImmutableInstalls false
+    - run: yarn
+    - run: yarn lint
+    - run: yarn build
     - run: npm publish
       env:
         NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
