+++ .yarnrc.yml
@@ -1,0 +1,3 @@
+compressionLevel: mixed
+
+nodeLinker: node-modules
